Used Tk::AstroCatalog instead of Tk::SrcCatalog.
Tk::AstroCatalog is part of the Astro::Catalog repository.

When sourceplot opens a catalog window, it specifies the correct
file name and instructs Astro::Catalog::IO::JCMT not to add
planets to the end of the list.  However this does not
alter the behavior of the "Default Catalog" menu option within
Tk::AstroCatalog, which will invoke the default loading options
of that class.  This adds planets and uses a fixed location for the
catalog file.  The module doesn't appear to offer a way to change
this except via the ASTRO_CATALOG_JCMT environment variable.