Open new source window with blank fields.
Change requested by Holly Thomas.

When the "New" button is clicked to open the new source
window, the RA and Dec fields are now empty and RJ is
selected.  This restores the classic behaviour of
SourcePlot and is more useful as RJ is more commonly
required than AZ and it saves the user from having to
delete the default coordinates from the boxes.