Simplify calcTime function
It looks like the second part of the if ... else ... construct
was never being used (source specified with timeDotX already
filled in).  Make a list of sources to plot -- either the
given source or the complete list -- rather than always iterating
over the complete list and matching the given source.

Also specify plotting style and relative size of the time
dot as arguments to the plot_time_dot function.