Add option to try to plot JCMT pointing position
This adds a command line option "--telposn" to the sourceplot script
which can be used to specify a telescope.  (Currently only JCMT is
supported.)  When given this option, the script should construct an
object of the new class App::SourcePlot::TelPosn::JCMT which has a
method "get_position" to get an App::SourcePlot::Source object
representing the current pointing position.  The calcTime method
applies this when plotting all sources, or when given the string
"TelescopePosition" in place of a source object.

Note: this implementation causes the "caget" command to be run each
time calcTime is scheduled (typically every 30 seconds and on updates
to the plot).  If this turns out not to be a good idea then a different
method of obtaining the current position will be required.