Add SAMP support
In the absence of a usable SAMP library for Perl, this currently
works by running a copy of StarJava's JSamp snooper tool.  It listens
for the mtype table.load.votable and should add any coordinates
received to the source list.  Note that not all varieties of
VOTable will be supported, but the hope is to be able to read
tables sent by the JCMT Query Tool.